Cythonized the computation of the magnetization for microncanonical chain. Nevertheless, the computation is much faster with the 'old' broadcasting for the sum, than with the cython loop.
Therefore we are still using the broadcasting (should a pure C function
be written? probably). In order to avoid overflow problems, one can take
the log, but then it is not possible to use a sum in the broadcasting. We
choose instead to normalize the probability (free entropy) so that it is
not too large.